pvr: Support semi-transparent pixels when blending textures
Texture blending is done conditionally: semi-transparent pixels (those
with the bit 15 set) are blended, while regular transparent and opaque
pixels are not blended.

Create a mask of (semi-transparent OR transparent) pixels into a
separate 4bpp texture, which will be used during the blending step to
make sure that only the semi-transparent pixels are blended with the
background.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>