pvr: Use direct rendering... directly
Do not use pvr_dr_init(), so that KOS won't try to lock the SQs for us.
We don't want that, because we cannot have the SQs locked at all times,
as doing so disables the MMU.

This is fine, as we can still call pvr_dr_target() on a unsigned int
variable; the only thing it will do is flip the 5th bit of the word at
each call.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>