pvr: Fix line rendering
The poly context did not have the right bits set. It enabled
alpha-blending for the vertex colors, which broke as soon as the
blending code started to support transparency support. It also did not
set the culling and more importantly the Z-buffer parameters, which is
important for the mask bit emulation.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>