pvr: Make sure the dest alpha channel is reset after blending
Update the B+F blending, as well as the code that applies the mask
texture, to make sure that they will always reset the alpha channel of
the accumulation buffer back to 0xff on every pixel.

This is necesary because the blending algorithms depend on the
accumulation buffer's alpha channel to be 0xff always.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>