pvr: Handle color-blending commands with opaque textures
Tag the cached textures with an 'has_semi' flag, set when there are some
semi-transparent texels in the texture. Without this flag, when no
texture mask was passed, we cannot know if the texture has no
semi-transparent texels, or has only semi-transparent or transparent
texels.

If we render a textured polygon or rectangle with alpha blending, and
the 'has_semi' flag is false, none of the texels will alpha-blend - we
can then handle the command as if alpha blending was disabled.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>