pvr: Support "brighter than bright" colors
When rendering textured blended polys and rectangles, the brightest
colors are 0x80; values above that are "brighter than bright",
allowing the textures to be rendered up to twice as bright as how
they are stored in memory.

If each subpixel is below that threshold, we can simply double the
vertex color values. Otherwise, we have to handle the brighter pixel
colors in the blending routine.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>