cdr: Fix reading of subchannel data
The BIOS syscall seems to read two more bytes than requested, which
causes a buffer overflow. Alternatively, it is also possible that the
syscall parameter may not be the size of the buffer in bytes.

Increase the subchannel data buffer from 100 bytes (4-byte header + 96
bytes of subchannel data) to 102 bytes to account for those two extra
bytes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>