pvr: Delay destruction of textures
It is not safe to destroy the textures as soon as the renderer is done
with them, because the PVR might still use them for rendering; if that's
the case, overwriting the old texture's memory space with new data would
cause graphical glitches.

Address this issue by enqueueing the stale textures to a reap list,
that's emptied when we know for sure that the PVR won't use them
anymore.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>