pvr: Populate codebooks in lookup order
Previously, the code would try to find a codebook for a given palette by
parsing the palette array in increasing order, and if not found, would
use the last free spot to register the palette.

This meant that the palettes were added into the array in reverse order,
from the end to the start, causing the codebook lookup to waste a lot of
time finding the codebooks.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>