pvr: Handle invalidation of individual subtexture blocks
When the CPU writes an area in VRAM, even a tiny area, we would
previously toss away entire pages, even when the invalidated area did
not touch any loaded subtexture blocks.

Improve things by simply tossing away the indidual subtexture blocks
covered by the invalidation area.

Make sure that if a particular subtexture block has already been used
for the frame, we create a new texture page, so that the old subtexture
is still available (in the stashed texture page) when the rendering is
actually done.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>