pvr: Fix and simplify B-F blending
B-F blending has been broken since the big texture overhaul change.

I'm not sure what was the cause, but it seems that the alpha bits in the
accumulation buffer were not guaranteed to be 0xff anymore - probably
because the mask texture is unconditionally applied now.

Address this issue by reworking the B-F blending so that the alpha bits
always stay 0xff.

The previous code was inverting B, then additive-blending F with
"srcalpha" source, so that only non-transparent pixels were copied. This
caused the alpha bits to be all over the place in the accumulation
buffer, so those were cleared to 0x00 afterwards. In turn, this required
a different algorithm to load and apply the mask texture, and required
the alpha bits to all be reset to 0xff in the end.

The new code is smarter; it still inverts B, but then unconditionally
does additive-blending with the foreground, with the foreground's alpha
bits set to 0x00. This works, because additive-blending a transparent
pixel is the same as not blending it. The consequence is that the alpha
bits all stay at 0x00, and the second invert reset them to 0xff.

Therefore the alpha bits are now always 0xff, and we can drop the
special case for loading the mask texture with subtractive-blending.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>