pvr: Change how mask textures work
Instead of having a mask texture composed of 0xffff for transparent or
semi-transparent pixels and 0x0000 for opaque pixels, and use some
blending trickery to apply the mask, store directly opaque
non-semi-transparent pixels in the mask texture (which isn't really a
mask anymore).

The advantage of this approach, is that the mask texture can now be
simply applied on top of the blend polygons. Even better, in the future
it can be rendered as a Punch-Through polygon, which will cause anything
behind the opaque pixels to not be rendered.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>