pvr: Use Punch-Through polygons when possible
Punch-Through polygons are much faster to render than transparent
polygons. They also have another advantage: any transparent polygons
behind opaque pixels of a punch-through polygon won't be rendered,
saving precious rendering time.

One drawback is that the depth comparison of punch-through polygons is
fixed as "greater or equal". This makes handling the mask bit a bit
tricky. As soon as the "set bit" is *cleared*, we have to switch from
using punch-through to using transparent polygons unconditionally so
that the rendering order is respected.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>