pvr: Handle texture mask as regular palette
Masks are actually just different palettes, so we can treat them as
regular palettes. Use bit 15 of CLUTs as the marker to differenciate
them.

The advantage of handling masks as regular palettes, is that we do not
need to store one mask per palette anymore in the codebooks, and the
codebook sizes can later be halved, meaning we will get twice as many.
As only a few CLUTs will need a corresponding mask, this will cause
a nice gain of space inside the codebooks areas.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>