CMake: Change CODE_BUFFER_SIZE to CODE_BUFFER_SIZE_MB
The emulator uses 1 MiB pages to map the emulated RAM, and expect the
dedicated memory area to be located at the end of the Dreamcast's RAM.
When configured for anything not megabyte-aligned, the page mapping
would silently fail, causing all kind of issues later on.

Address this issue by making it impossible to pick a code buffer size
that's not properly aligned.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>