CMake: Add option WITH_HYBRID_RENDERING
Add an option to enable hybrid rendering, and leave it disabled by
default.

Hybrid rendering means that two PVR lists are used to submit the
geometry, which generally allows for faster rendering. However, it is
not yet working perfectly, as some games will overflow the vertex buffer
used for the deferred list.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>