pvr: Rework find_texture_codebook()
Avoid using the 15th bit of CLUTs as an 'entry valid' bit. Since we
always only allocate entries in CLUT arrays in increasing linear order,
and never deallocate them individually, we can just keep a counter to
know how many entries are valid and how many are available. This has
the added benefit that we do not have to parse the whole CLUT array to
find out a specific CLUT.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>