pvr: Reduce size of codebooks to increment their count
Reduce the size of a 4bpp codebook from 512 to 256 bytes, and the size
of a 8bpp codebook from 4096 to 2048 bytes, by removing the useless mask
field (+ padding). Also reduce the padding after 4bpp codebooks from
2048 bytes to 1792 bytes to fit another 256-byte 4bpp codebook.

This means that we can now store 249 4bpp codebooks instead of 124
previously, and 32 8bpp codebooks instead of 16 previously.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>