pvr: Experimental palette invalidation mechanism
Previously, palette invalidation did not really work, as there was no
mechanism in place to detect whether or not a palette contained data
that was still valid and hadn't been invalidated.

This only sort-of worked by chance, as the palette used with a given
texture page are commonly found inside that same texture page.

But this also caused some other games to show a lot of graphical
glitches.

Solve this issue by keeping track of the number of times any texture
page is invalidated, as an "invalidation counter" in the PVR struct, and
in each and every texture page. When a texture page is invalidated, its
counter is reset to the one in the PVR struct.

When creating a palette, the counter of the texture page it's located in
will be memorized. When trying to use the palette, it will be compared
vs. the counter of that same page; if it does not match, the palette is
stale and should be invalidated.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>