Implement saving to / loading from VMU
Implement memcard hot-plug. By default, the first memcard corresponds to
the VMU in A1, the second memcard corresponds to the VMU in B1.

When a VMU is present at boot, or plugged in, a gzip-compressed file is
loaded into the in-RAM memcard buffer. If the file does not exist, the
in-RAM memcard buffer is instead initialized as freshly formatted, and
the VMU file will be written the first time the file is opened for
writing.

The code implements a VFS, which allows the emulator to read/write the
memcard image as if it was uncompressed, while keeping it compressed on
the VMU. This is because an uncompressed memcard image is 128 KiB, which
is the exact size of a VMU.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>