pvr: Add struct poly + helper functions
This "poly" struct contains all the information needed for the PVR
plugin to render a PSX polygon.

The point is that at only 64 bytes, it is much smaller than the
equivalent PVR polygon in a DMA buffer, which goes from 128 bytes for a
3-vertices polygon (with the header included) to 160 bytes for a
4-vertices polygon. It is also conveniently aligned to two cache lines.

The driver will then be able to use direct-rendering for the polygons
that have to go through the currently opened list, and queue the other
ones as "struct poly", to be processed (not DMA'd) later when their
respective list has been opened.

The hope is that it will work just as good as hybrid (DMA) rendering,
without requiring as much RAM.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>