pvr: Re-introduce deferred rendering
Default to sending everything to the punch-thru list when possible.
Otherwise, the polys that need to be rendered on the transparent list
will have their poly struct copied into a temporary buffer.

Once the renderer is done parsing the PSX commands for the current
frame, the punch-thru list will be closed, the transparent list will be
opened, and and all polys stored in the temporary buffer will then be
processed for rendering.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>