pvr: Include draw offsets in poly coordinates
If we want to be able to defer rendering of a given poly, we cannot
assume that the current draw offsets will be the same at the time where
we will render the poly. Therefore, the draw offsets should be added to
the X/Y coordinates of the poly when the "struct poly" is created, and
the result can then be translated to screen coordinates during
rendering.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>