pvr: Compute and apply vertical offset later
Commit 8568cb62 ("pvr: Apply vertical offset early") moved the vertical
offset computation so that it's done as early as possible.

This made sense at that point, but it is now a problem: it requires the
codebook to be known, which means the texture page needs to be known.
This does not really work well with the indirect rendering mechanism,
where the texture page and codebook should be computed right before
drawing, for reasons explained in the next commit.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>