pvr: Move check for new frame
The check for new frame (to know when to start a scene) was in
draw_prim(), which is called very often, and was therefore more costly
than it should have been.

Move it to poly_enqueue(), when the polys are drawn directly. This also
means that this check won't be done twice for the enqueued polys when
doing hybrid rendering.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>