pvr: Enhance CLUT invalidation
Color palettes for 8bpp frames can cover more than one texture page.
Therefore we need to check the inval counter of each covered page to
know if the CLUT is outdated.

When we find our CLUT in the CLUT array, but it is outdated, check if it
was used for the current frame. If it is the case, we have to use
another codebook. Otherwise, we can simply reload the palette.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>