pvr: Support inval counter warping around
We cannot do a direct comparison (a >= b) of two timer values a/b,
because those are uint16_t and will warp back to 0 at some time, in
which case the comparison will be invalid.

What we can do, is compare them relative to the current value of the
main inval counter.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>