pvr: Pass vertex count information as a flag
The vertex count is either 3, or 4. Instead of using a uint8_t value to
carry this information in the poly structure, we can carry the
information as a single bit.

This might help optimization (that's a supposition) as the compiler now
knows that the value can either be 3 or 4 and can act accordingly.

More importantly, it frees up some space in the poly structure which
will be reused later.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>