pvr: Store texture and voffset in poly struct
Getting a handle to the texture and computing the voffset at the time of
drawing works only when rendering to the primary list. For everything
that's deferred we might not get the correct texture, if the texture has
been invalidated after we queued the poly.

Solve this issue by using the newly created space inside the poly
structure to store a texture pointer and the voffset, and compute those
at the time of queueing the blocks.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>