Rework exit mechanism
The reset key combo (START + A + B + X + Y) will now return to the main
menu, instead of exiting the emulator. There, a new ISO can be selected,
or a new PSX disc can be used. To exit the emulator, the "Quit" option
of the main menu should now be used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>