pvr: Add support for overpaint
When the render window in VRAM is the same as the view window, the PSX
program is rendering in single-buffering mode, which means that it will
paint over the frame that's currently being displayed.

In that case, most often than not it will only paint over a subset of
the view window, which means that some of the pixel data of the previous
frame will carry on to the new frame. This is visible right at the BIOS
screen, for instance, when the PlayStation logo appears.

Emulate overpaint by loading the front buffer as a texture to a
screen-sized quad placed on the background.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>