pvr: Don't open TR list for overpaint when we won't overpaint
When the TR list is not opened during the rendering of the scene, we
need to open it before finishing rendering, in order to render the front
buffer for overpaint.

However, we only need to do that when overpaint is needed... otherwise
the list will be just opened then closed without any poly being
submitted.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>