pvr: Fix coordinates once again
We can't rely on pvr.start_x / pvr.start_y to have their final value
when updating the drawing offsets. Therefore we need to compute a new
pixel offset that's the sum of the GTE drawing offset + start offset,
and that gets updated when any of those two gets updated.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>