input: Scale analog inputs
The analog sticks on the PSX behave differently than the sticks on the
Dreamcast (or most other platforms). If we plot the X/Y values while
doing circular motions on the sticks, on the PSX we would be drawing a
square, while on the Dreamcast we would be drawing a circle.

The PSX games are designed with that in mind, which means that they
expect the joystick's X/Y values to go to their maximum when aiming
towards the top-right. Support that by scaling the analog input values
by a fixed factor.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>