Set vertical scale to 1.0x on non-VGA
The vertical scale (used when rendering the final image to the
framebuffer) defaults to 1.0x on VGA, but only 0.999x on non-VGA, to
apply a smoothing effect.

This does not work well with the framebuffer reload feature needed for
single-buffered rendering. Force a vertical scale of 1.0x to address
this issue.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>