pvr: Cut polys overlapping texture pages into smaller polys
One problem with 8bpp and 16bpp textures, is that their texture pages
overlap. This means that we currently use way more VRAM than we should.

One solution to this is to use smaller texture pages: with 128x256 pages
for 8bpp and 64x256 pages for 16bpp, there is no overlap.

The problem then is that the PSX can render polygons that need the full
256x256 texture page. To make it work with smaller texture pages, cut
the polygons that cover more than one texture page into smaller polys.

After this change, the texture pages will still be 256x256 but the polys
will only use the first 64x256 for 16bpp or the first 128x256 for 8bpp.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>