pvr: Reduce 16bpp pages to 64x256 and 8bpp pages to 128x256
Now that our 16bpp and 8bpp textured polys are scissored to a max width
of 64 half-words, we don't actually need to allocate texture pages big
enough to accomodate 256x256 pixels.

Update the code so that we allocate 64x256 16bpp pages and 128x256 8bpp
pages, saving a lot of VRAM in the process.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>