pvr: Increase vertical size of fake-VQ textures
The texture pages are 256x256, but since we're using fake VQs, those are
expanded to 512x512.

Fake VQs work by abusing the VQ format to store multiple palettes
alongside a texture. By computing a fake-VQ address from the address of
the frame and the index of the palette that we want, and adjusting the V
coordinate accordingly, we can render the frame with different palettes.

The problem then is that this offset added to the V coordinate can push
the actual V coordinate above 1.0, in which case the texture will start
to flip around.

Address this issue by lying about our texture being 1024 lines in total,
and adjusting the V scale factor accordingly.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>