pvr: Update subtexture block sizes
Instead of having subtexture blocks of 8x8 pixels, use subtexture blocks
of 16 half-words per 16 lines. This means that the width of a block will
always be 32 bytes, which is perfect for SQ transfers.

Given that our texture pages now have a max. size of 64 half-words per
256 lines, this also makes sure that we fully utilize the 64 bits of
the subtexture block masks.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>