pvr: Support loading 16-bit background from emulated VRAM
When the emulated PSX game copies data to the back buffer in the
emulated VRAM, load the copied data as a textured background into the
current scene.

The actual loading of the data to PVR textures is delayed until the end
of the scene, as some games (notably Rayman) seem to like to overpaint.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>