pvr: Speed up loading of subtexture blocks
Instead of locking SQs before each block then unlocking SQs once that
block is transferred, lock the SQs before processing a texture page, and
unlock the SQs once all the subtexture blocks of the texture page have
been updated.

Also, 16bpp textures are now loaded much faster as two pixels are
processed at a time, and loaded directly to the SQs instead of being
loaded to an intermediary buffer that's then copied to the SQs.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>