pvr: Working pixel clipping
During the scene the game may change the render area a few times.
For each change, render a modifier volume as a rectangular cuboid
whose X/Y coordinates delimitate the render area, and the Z
coordinates deliminate the start and end depth of the render area.
Those volumes are then rendered as "exclude" modifiers, and an
"include" modifier plane is rendered on top. The result is that only
pixels inside those volumes will be rendered, anything outside will
be clipped.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>