Re-enable list auto-sort
Disabling auto-sort permitted to use the "always" depth comparison,
which was needed for the sticky bit emulation, back when sticky bit
emulation was done using the Z buffer.

Now that it's emulated using the destination alpha bits, we don't use
anything else than the "greater or equal" depth comparison, we can
re-enable list auto-sort.

Note that enabling list auto-sort is supposed to carry a performance
penalty when many transparent polygons are used. On the other hand, it
allows to use volume modifiers with TR polygons, which will later be
used to implement pixel clipping on the PVR.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>