pvr: Fix mask not applied when checking for sticky bits
The original idea was that if we're checking the mask, it means that we
sent a TR poly, so the mask would also be a TR poly and not a PT poly,
which wouldn't cause the PVR to discard the pixels of the first poly.

This is true, but we still need to re-apply the opaque pixels using the
mask as the bottom pixels will be unconditionally blended.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>