input: Consider all pads as analog pads
Regular Dreamcast controllers have an analog pad, so it makes sense to
report a dual shock. Besides, if we report a digital pad, the games
won't try to use rumble.

This seems to work fine even with games that require a digital pad, as
PCSX will automatically toggle the analog mode.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>